关于数据最大值U32的处理

+++ ssd_interface.c
@@ -2359,7 +2359,7 @@ int ADFTL_Find_Victim_In_RCMT_W()
 
     int *index_arr,*lpn_arr,*Time_arr;
     //32位最大的int值
-    int MAXTIME=21474836473;
+    long  MAXTIME=21474836473;
     //debug
     if(operation_time>=MAXTIME){
         printf("operation_time is %d,over MAXTIME\n",operation_time);
@@ -2369,7 +2369,7 @@ int ADFTL_Find_Victim_In_RCMT_W()
     //存的是real_arr的下标索引
     index_arr=(int *)malloc(ADFTL_WINDOW_SIZE*sizeof(int));
     lpn_arr=(int *)malloc(ADFTL_WINDOW_SIZE*sizeof(int));
-    Time_arr=(int *)malloc(ADFTL_WINDOW_SIZE*sizeof(int));
+    Time_arr=(long *)malloc(ADFTL_WINDOW_SIZE*sizeof(long));
     //init
     for(i=0;i<ADFTL_WINDOW_SIZE;i++){
         index_arr[i]=-1;
